a2a/airline/pyproject.toml
[project]
name = "airline"
version = "0.1.0"
description = "Airline booking agent with A2A server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents[a2a]",
    "bedrock-agentcore",
    "openai",
]
